// src/system/types.ts
export type MessagePage = 'background' | `content/${typeof MODULE_NAME}` | 'popup' | 'options';

interface MessageErrorResponse { error: true; message: string }
interface MessageSuccessResponse<T> { error: false; data: T }

export type MessageResponse<T> = MessageErrorResponse | MessageSuccessResponse<T>;

export type MessageListener<T = any, TReturn = any> = (
    data: T,
    meta: {
        sender: chrome.runtime.MessageSender;
        ok: (data: TReturn) => MessageSuccessResponse<TReturn>;
        fail: (message: string) => MessageErrorResponse;
    },
) => MessageResponse<TReturn> | Promise<MessageResponse<TReturn>> | void | Promise<void>;

export type Message<T = any> = {
    namespace: undefined;
} | {
    namespace: string;
    targetPage: MessagePage;
    name: string;
    data?: T;
};